+++ mailbag/derivatives/pdf.py
@@ -48,9 +48,18 @@ if not skip_registry:
             pdf_name = filename + ".pdf"
 
             if message.HTML_Body is None and message.Text_Body is None:
-                log.warn("No HTML or plain text body for " + str(message.Mailbag_Message_ID) + ". No PDF derivative will be created.")
+                log.warn(
+                    "No HTML or plain text body for "
+                    + str(message.Mailbag_Message_ID)
+                    + ". No PDF derivative will be created."
+                )
             else:
-                log.debug("Writing HTML to " + str(html_name) + " and converting to " + str(pdf_name))
+                log.debug(
+                    "Writing HTML to "
+                    + str(html_name)
+                    + " and converting to "
+                    + str(pdf_name)
+                )
                 html_formatted, encoding = helper.htmlFormatting(message, self.args.css)
 
                 if not self.args.dry_run:
@@ -60,18 +69,40 @@ if not skip_registry:
                     with open(html_name, "w", encoding=encoding) as write_html:
                         write_html.write(html_formatted)
                         write_html.close()
+                    command = [
+                        wkhtmltopdf,
+                        "--disable-javascript",
+                        os.path.abspath(html_name),
+                        os.path.abspath(pdf_name),
+                    ]
+                    log.debug("Running " + " ".join(command))
                     p = subprocess.Popen(
-                        [wkhtmltopdf, "--disable-javascript", html_name, pdf_name], stdout=subprocess.PIPE, stderr=subprocess.PIPE
+                        command, stdout=subprocess.PIPE, stderr=subprocess.PIPE
                     )
                     stdout, stderr = p.communicate()
                     if p.returncode == 0:
-                        log.debug("Successfully created " + str(message.Mailbag_Message_ID) + ".pdf")
+                        log.debug(
+                            "Successfully created "
+                            + str(message.Mailbag_Message_ID)
+                            + ".pdf"
+                        )
                     else:
                         if stdout:
-                            log.warn("Output converting to " + str(message.Mailbag_Message_ID) + ".pdf: " + str(stdout))
+                            log.warn(
+                                "Output converting to "
+                                + str(message.Mailbag_Message_ID)
+                                + ".pdf: "
+                                + str(stdout)
+                            )
                         if stderr:
-                            log.error("Error converting to " + str(message.Mailbag_Message_ID) + ".pdf: " + str(stderr))
+                            log.error(
+                                "Error converting to "
+                                + str(message.Mailbag_Message_ID)
+                                + ".pdf: "
+                                + str(stderr)
+                            )
                         # comment out for now since catches even very minor issues
                         # raise TypeError(stderr)
                     # delete the HTML file
-                    os.remove(html_name)
+                    if os.path.isfile(pdf_name):
+                        os.remove(html_name)
